requirements.txt
fastapi==0.104.0
uvicorn==0.23.2

torch==2.1.0
torchvision==0.16.0
numpy<2
pillow==10.1.0

python-multipart==0.0.6

scikit-learn>=1.3